Support Streams bigger than INT_MAX
In Dwrf/ORC streams can be bigger than size INT_MAX (2GB)
though they are rare, they are possible.
This change introduces support for size bigger than 2GB.

It also adds a checksum file to compute checksum given any DWRF file.

+++ presto-orc/src/main/java/com/facebook/presto/orc/DiskRange.java
@@ -23,9 +23,9 @@ import static java.util.Objects.requireNonNull;
 public final class DiskRange
 {
     private final long offset;
-    private final int length;
+    private final long length;
 
-    public DiskRange(long offset, int length)
+    public DiskRange(long offset, long length)
     {
         checkArgument(offset >= 0, "offset is negative");
         checkArgument(length > 0, "length must be at least 1");
@@ -39,7 +39,7 @@ public final class DiskRange
         return offset;
     }
 
-    public int getLength()
+    public long getLength()
     {
         return length;
     }
